requirements.txt
TwitterAPI
python-twitter
boto3
